+++ forge-runtime/src/main/java/com/kivojenko/spring/forge/jpa/model/relation/manyToOne/ReadManyToOneEndpointRelation.java
@@ -1,0 +1,29 @@
+package com.kivojenko.spring.forge.jpa.model.relation.manyToOne;
+
+import com.kivojenko.spring.forge.jpa.model.relation.EndpointRelation;
+import com.squareup.javapoet.MethodSpec;
+import lombok.experimental.SuperBuilder;
+
+import javax.lang.model.element.Modifier;
+
+import static com.kivojenko.spring.forge.jpa.generator.MethodGenerator.GET_MAPPING;
+
+/**
+ * Represents a relation that generates a GET endpoint to read a Many-to-One associated entity.
+ */
+@SuperBuilder
+public class ReadManyToOneEndpointRelation extends EndpointRelation {
+    @Override
+    public MethodSpec getControllerMethod() {
+        return MethodSpec
+                .methodBuilder(methodName)
+                .addModifiers(Modifier.PUBLIC)
+                .addAnnotation(annotation(GET_MAPPING))
+                .returns(targetEntityModel.entityType())
+                .addParameter(baseParamSpec())
+                .addStatement("return getById($L).$L()", BASE_ID_PARAM_NAME, methodName)
+                .build();
+    }
+
+}
+