+++ forge-runtime/src/main/java/com/kivojenko/spring/forge/jpa/model/relation/oneToMany/OneToManyEndpointRelation.java
@@ -1,0 +1,43 @@
+package com.kivojenko.spring.forge.jpa.model.relation.oneToMany;
+
+import com.kivojenko.spring.forge.jpa.model.relation.ServiceRepositoryEndpointRelation;
+import com.squareup.javapoet.MethodSpec;
+import jakarta.persistence.EntityNotFoundException;
+import lombok.experimental.SuperBuilder;
+
+import static com.kivojenko.spring.forge.jpa.generator.MethodGenerator.getterName;
+import static com.kivojenko.spring.forge.jpa.utils.StringUtils.decapitalize;
+
+
+/**
+ * Base class for One-to-Many endpoint relations.
+ */
+@SuperBuilder
+public abstract class OneToManyEndpointRelation extends ServiceRepositoryEndpointRelation {
+    protected String mappedBy;
+
+    protected void addFindSub(MethodSpec.Builder methodSpec) {
+        methodSpec
+                .addParameter(subParamSpec())
+                .addStatement(
+                        "var $L = $L.findById($L).orElseThrow($T::new)",
+                        SUB_VAR_NAME,
+                        decapitalize(targetEntityModel.repositoryName()),
+                        SUB_ID_PARAM_NAME,
+                        EntityNotFoundException.class
+                )
+                .beginControlFlow(
+                        "if (!$N.$L().$L().equals($L))",
+                        SUB_VAR_NAME,
+                        getterName(mappedBy),
+                        getterName(entityModel),
+                        BASE_ID_PARAM_NAME
+                )
+                .addStatement("throw new $T($S)",
+                        IllegalStateException.class,
+                        entityModel.entityType().simpleName() +
+                                " is not associated with the given " +
+                                targetEntityModel.entityType().simpleName())
+                .endControlFlow();
+    }
+}